shg_os_app/requirements.txt
streamlit
pandas
numpy
plotly
google-generativeai
python-dotenv
Pillow
requests
streamlit
pandas
numpy
matplotlib
Pillow
python-dotenv
sqlite-utils
google-generativeai
protobuf